git-going-backend/requirements.txt
brotlipy==0.7.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
Deprecated==1.2.7
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.2
pycares==3.1.0
pycparser==2.19
PyGithub==1.44.1
PyJWT==1.7.1
pymongo==3.10.0
requests==2.22.0
retrying==1.3.3
six==1.13.0
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.4.2